# modulo-3/semana-8/exercicios/10/swagger.yaml
swagger: '2.0'
info:
  description: API
  version: '1.0'
  title: API do grupo 05
  license:
    name: Apache License Version 2.0
    url: http://teste.com
host: 'localhost:8080'
basePath: /
tags:
  - name: assunto-controller
    description: Assunto Controller
  - name: interessado-controller
    description: Interessado Controller
  - name: processo-controller
    description: Processo Controller
paths:
  /v1/assuntos:
    get:
      tags:
        - assunto-controller
      summary: findAll
      operationId: findAllUsingGET
      produces:
        - application/json
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/AssuntoDTOOutput'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - assunto-controller
      summary: create
      operationId: createUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: assuntoDTO
          description: assuntoDTO
          required: true
          schema:
            $ref: '#/definitions/AssuntoDTOInput'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/AssuntoDTOOutput'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/v1/assuntos/{id}':
    get:
      tags:
        - assunto-controller
      summary: find
      operationId: findUsingGET
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/AssuntoDTOOutput'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - assunto-controller
      summary: update
      operationId: updateUsingPUT
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: assuntoDTO
          description: assuntoDTO
          required: true
          schema:
            $ref: '#/definitions/AssuntoDTOInput'
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '201':
          description: Created
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - assunto-controller
      summary: delete
      operationId: deleteUsingDELETE
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /v1/interessados:
    get:
      tags:
        - interessado-controller
      summary: findAll
      operationId: findAllUsingGET_1
      produces:
        - application/json
      parameters:
        - name: nu_identificacao
          in: query
          description: nu_identificacao
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/InteressadoDTOOutput'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - interessado-controller
      summary: create
      operationId: createUsingPOST_1
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: interessadoDTO
          description: interessadoDTO
          required: true
          schema:
            $ref: '#/definitions/InteressadoDTOInput'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/InteressadoDTOOutput'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/v1/interessados/{id}':
    get:
      tags:
        - interessado-controller
      summary: find
      operationId: findUsingGET_1
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/InteressadoDTOOutput'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - interessado-controller
      summary: update
      operationId: updateUsingPUT_1
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
        - in: body
          name: interessadoDTO
          description: interessadoDTO
          required: true
          schema:
            $ref: '#/definitions/InteressadoDTOInput'
      responses:
        '201':
          description: Created
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - interessado-controller
      summary: delete
      operationId: deleteUsingDELETE_1
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /v1/processos:
    get:
      tags:
        - processo-controller
      summary: findAll
      operationId: findAllUsingGET_2
      produces:
        - application/json
      parameters:
        - name: cd_assunto_id
          in: query
          description: cd_assunto_id
          required: false
          type: integer
          format: int64
        - name: cd_interessado_id
          in: query
          description: cd_interessado_id
          required: false
          type: integer
          format: int64
        - name: chave_processo
          in: query
          description: chave_processo
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/ProcessoDTOOutput'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - processo-controller
      summary: create
      operationId: createUsingPOST_2
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: processoDTO
          description: processoDTO
          required: true
          schema:
            $ref: '#/definitions/ProcessoDTOInput'
      responses:
        '201':
          description: Created
          schema:
            $ref: '#/definitions/ProcessoDTOOutput'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  '/v1/processos/{id}':
    get:
      tags:
        - processo-controller
      summary: find
      operationId: findUsingGET_2
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/ProcessoDTOOutput'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    put:
      tags:
        - processo-controller
      summary: update
      operationId: updateUsingPUT_2
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
        - in: body
          name: processoDTO
          description: processoDTO
          required: true
          schema:
            $ref: '#/definitions/ProcessoDTOInput'
      responses:
        '201':
          description: Created
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    delete:
      tags:
        - processo-controller
      summary: delete
      operationId: deleteUsingDELETE_2
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
      responses:
        '204':
          description: No Content
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
definitions:
  Assunto:
    type: object
    properties:
      descricao:
        type: string
      dtCadastro:
        type: string
        format: date
      flAtivo:
        type: string
      id:
        type: integer
        format: int64
    title: Assunto
  AssuntoDTOInput:
    type: object
    required:
      - descricao
    properties:
      descricao:
        type: string
        minLength: 1
        maxLength: 250
      flAtivo:
        type: string
    title: AssuntoDTOInput
  AssuntoDTOOutput:
    type: object
    properties:
      descricao:
        type: string
      dtCadastro:
        type: string
        format: date
      flAtivo:
        type: string
      id:
        type: integer
        format: int64
    title: AssuntoDTOOutput
  Interessado:
    type: object
    properties:
      dtNascimento:
        type: string
        format: date
      flAtivo:
        type: string
      id:
        type: integer
        format: int64
      nmInteressado:
        type: string
      nuIdentificacao:
        type: string
    title: Interessado
  InteressadoDTOInput:
    type: object
    required:
      - dtNascimento
      - nmInteressado
      - nuIdentificacao
    properties:
      dtNascimento:
        type: string
        pattern: >-
          ^(?:(?:31(\/|-|\.)(?:0?[13578]|1[02]))\1|(?:(?:29|30)(\/|-|\.)(?:0?[13-9]|1[0-2])\2))(?:(?:1[6-9]|[2-9]\d)?\d{2})$|^(?:29(\/|-|\.)0?2\3(?:(?:(?:1[6-9]|[2-9]\d)?(?:0[48]|[2468][048]|[13579][26])|(?:(?:16|[2468][048]|[3579][26])00))))$|^(?:0?[1-9]|1\d|2[0-8])(\/|-|\.)(?:(?:0?[1-9])|(?:1[0-2]))\4(?:(?:1[6-9]|[2-9]\d)?\d{2})$
      flAtivo:
        type: string
      nmInteressado:
        type: string
        minLength: 1
        maxLength: 250
      nuIdentificacao:
        type: string
    title: InteressadoDTOInput
  InteressadoDTOOutput:
    type: object
    properties:
      dtNascimento:
        type: string
        format: date
      flAtivo:
        type: string
      id:
        type: integer
        format: int64
      nmInteressado:
        type: string
      nuIdentificacao:
        type: string
    title: InteressadoDTOOutput
  ProcessoDTOInput:
    type: object
    required:
      - cdAssuntoId
      - cdInteressadoId
      - descricao
      - nuAno
      - sgOrgaoSetor
    properties:
      cdAssuntoId:
        type: integer
        format: int64
      cdInteressadoId:
        type: integer
        format: int64
      descricao:
        type: string
        minLength: 1
        maxLength: 250
      nuAno:
        type: string
        minLength: 1
        maxLength: 4
      sgOrgaoSetor:
        type: string
        minLength: 1
        maxLength: 4
    title: ProcessoDTOInput
  ProcessoDTOOutput:
    type: object
    properties:
      cdAssunto:
        $ref: '#/definitions/Assunto'
      cdInteressado:
        $ref: '#/definitions/Interessado'
      chaveProcesso:
        type: string
      descricao:
        type: string
      id:
        type: integer
        format: int64
      nuAno:
        type: string
      nuProcesso:
        type: integer
        format: int64
      sgOrgaoSetor:
        type: string
    title: ProcessoDTOOutput